Refactor LLM module structure and add SummarizeHandler
- Moved `SummarizeHandler` to a new `summarize.py` file within the `llm` directory for better organization.
- Introduced `LLMHandler` in `chatbot.py` to manage interactions with the OpenAI API.
- Updated imports in `yt_fts.py` to reflect the new module structure.
- Added functionality for video summarization and transcript handling in the `SummarizeHandler` class.
